+++ Projects/Feature/HomeFeature/Sources/Assembly/HomeAssembly.swift
@@ -1,3 +1,4 @@
+import ConfirmationDialogFeature
 import MassageDomainInterface
 import MealDomainInterface
 import Moordinator
@@ -17,7 +18,8 @@ public final class HomeAssembly: Assembly {
                 fetchMealInfoUseCase: resolver.resolve(FetchMealInfoUseCase.self)!,
                 loadCurrentUserRoleUseCase: resolver.resolve(LoadCurrentUserRoleUseCase.self)!,
                 applySelfStudyUseCase: resolver.resolve(ApplySelfStudyUseCase.self)!,
-                applyMassageUseCase: resolver.resolve(ApplyMassageUseCase.self)!
+                applyMassageUseCase: resolver.resolve(ApplyMassageUseCase.self)!,
+                confirmationDialogFactory: resolver.resolve(ConfirmationDialogFactory.self)!
             )
         }
     }
